Move human input questions into Pulse view
